Added UI changes for individual call history page.
